pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intellicli"
version = "0.1.0"
description = "一个智能命令行助手，结合大型语言模型和动态任务规划"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "IntelliCLI Team", email = "lovemaojiu@gmail.com"},
]
keywords = ["ai", "cli", "llm", "assistant", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9.0",
    "PyYAML>=6.0",
    "requests>=2.28.0",
    "google-generativeai>=0.3.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/yourusername/IntelliCLI"
Documentation = "https://github.com/yourusername/IntelliCLI#readme"
Repository = "https://github.com/yourusername/IntelliCLI"
Issues = "https://github.com/yourusername/IntelliCLI/issues"

[project.scripts]
intellicli = "intellicli.cli:main"
icli = "intellicli.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["intellicli*"]
exclude = ["tests*", "docs*"]

[tool.setuptools.package-data]
intellicli = ["*.yaml", "*.yml", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short" 